501st_Secret_Notes/requirements.txt
flask
msal
azure-identity
azure-keyvault-secrets
